move load components into a single folder
