+++ src/components/Loaders/Notes.tsx
